LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/flex-flow-expected.txt


FAIL 'flex-flow' does not supported 'row' assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
PASS 'flex-flow' does not supported 'column wrap'
PASS 'flex-flow' does not supported 'row-reverse wrap-reverse'
